refactor: separate UI components in profile, tasks, and team pages
- Break down monolithic page files into smaller, specialized components
- Create dedicated directories for each page's components
- Extract reusable UI sections into individual widget files

This refactoring improves code maintainability by applying separation of
concerns. Each component now has single responsibility, making the codebase
easier to navigate, understand, and test. Components can now be reused and
updated independently across the application.

Affected modules:
- lib/pages/profile/
- lib/pages/team/
- lib/pages/tasks/
