+++ flutter_mobile_app/pubspec.yaml
@@ -34,10 +34,10 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
-  http: ^1.3.0
+  http: ^1.4.0
   flutter_dotenv: ^5.2.1
   intl: ^0.20.2
-  socket_io_client: ^3.1.1
+  socket_io_client: ^3.1.2
   flutter_local_notifications: ^19.2.1
